changed name of context provider to use across all website
